feat: Create configuration file
This patch allows a user to create a configuration file for copiepate.
This enables re-using configuration between multiple launches of the
tool.

Configuration settings are optionnal and override default settings.
Configuration file settings can be overriden by passing a command line
option.

Note: verbosity configuration settings is not parsed as the logger is
initialized before the configuration file being parsed. We might want to
allows this feature in the future by switching to a more powerful
logger.