feat: add encryption over the wire
This patch update the network protocol to allow encryption and
authentication of messages sent over the network.

A secret is now necessary to share between the client and the server.
That secret can either be passed using the `--secret` cli parameter of
in the `secret` config field.

In case the user doesn't want to generate a secret key, they can use the
`--insecure` cli parameter which will use the default key. `--insecure`
must be set on both server and client to work. This parameter is not
recommended as it is equivalent to plaintext transmission.

One minor caveat with the current protocol is that the message type is
not encrypted. Which mean that an attacker could see what type of
messages are sent over the network. However given that there is only 3
types of frames currently, this is not a huge privacy issue.

Future improvement could go in:
* Having a compile time state machine with transition to make sure
  client and server state transitions are valid
* Encrypt message type
* Auditing would need to be done to be more confident on security